to sqlalchemy

+++ src/services/tasks/service.py
@@ -6,8 +6,8 @@ from celery.signals import celeryd_init
 from sentry_sdk.integrations.celery import CeleryIntegration
 
 from src.core import config, db
-from src.services.auth import service as auth_service
 from src.services.auth import models as auth_models
+from src.services.auth import service as auth_service
 from src.services.preorder import tasks as preorders_tasks
 from src.services.sheets import models as sheets_models
 from src.services.sheets import service as sheets_service
@@ -58,7 +58,8 @@ def update_order(parser: dict, row_id: int, data: dict):
     with db.session_maker() as session:
         creds = auth_service.get_first_superuser_sync(session)
     parser_model = sheets_models.OrderSheetParseRead.model_validate(parser)
-    sheets_service.update_row_data(creds.google, parser_model, row_id, data)
+    if creds.google is not None:
+        sheets_service.update_row_data(creds.google, parser_model, row_id, data)
 
 
 @celery.task(name="create_booster")
@@ -66,7 +67,8 @@ def create_booster(parser: str, data: dict):
     with db.session_maker() as session:
         creds = auth_service.get_first_superuser_sync(session)
     parser_model = sheets_models.OrderSheetParseRead.model_validate_json(parser)
-    sheets_service.create_row_data(auth_models.UserReadSheets, creds.google, parser_model, data)
+    if creds.google is not None:
+        sheets_service.create_row_data(auth_models.UserReadSheets, creds.google, parser_model, data)
 
 
 @celery.task(name="create_or_update_booster")
@@ -74,7 +76,8 @@ def create_or_update_booster(parser: str, value: str, data: dict):
     with db.session_maker() as session:
         creds = auth_service.get_first_superuser_sync(session)
     parser_model = sheets_models.OrderSheetParseRead.model_validate_json(parser)
-    sheets_service.create_or_update_booster(creds.google, parser_model, value, data)
+    if creds.google is not None:
+        sheets_service.create_or_update_booster(creds.google, parser_model, value, data)
 
 
 @celery.task(name="delete_booster")
@@ -82,7 +85,8 @@ def delete_booster(parser: str, value: str):
     with db.session_maker() as session:
         creds = auth_service.get_first_superuser_sync(session)
     parser_model = sheets_models.OrderSheetParseRead.model_validate_json(parser)
-    sheets_service.delete_booster(creds.google, parser_model, value)
+    if creds.google is not None:
+        sheets_service.delete_booster(creds.google, parser_model, value)
 
 
 @celery.task(name="manage_preorders")
